DRY conditions / make sibling conditions logical AND
Conditions as currently implemented seem to be inconsistent in how they are handled
and a little confusing to read code-wise.

Without this patch it would be valid to declare an AND with a regex at the same level of
the condition hierarchy, but only the AND will execute. This is somewhat confusing and unnecessary.

Additionally, the code is weird to read as well.

This patch makes all statements at a given level execute, and have their result a logical AND of their
return values.

This also DRYs up the code and makes it easier to extend by not requiring each invocation of the condition
to run all possible checks. There may be a perf benefit here as well, though it is likely to be small.

This is technically a behavior change, since we don't short-circuit on the first compound statement encountered,
however, I couldn't find that behavior documented here https://www.elastic.co/guide/en/beats/filebeat/current/defining-processors.html,
and that behavior is surprising. I think there's a good case to be made that the current behavior is a bug..

Looking at the git history, especially here: 2d35d50a60adb5bc5b5f68d58677c2a5c16ae386 it looks like the original intent was for either an
equals statement to run, or a compound statement. Later, more compound statements were added, seemingly inadvertently giving us the current logic.

Another approach might be to require a compound statement for multiple leafs, but I think preserving the current behavior, and making it more consistent
is nice, and a logical AND is a sensible default, and is easier to read since it reduces nesting. We *should* document that.