Enhance mapval to handle arrays
Introduces initial support for arrays into mapval. This required a bit of an internals refactoring, making value validation less of a special case. Now, validating a value involves returning a full mapval.Results map which gets merged into the main Validator results. This is a better internal design and allows for more flexibility.

A test was added for comparing array literals. This isn't something that's actually needed yet, but I wanted a test for the current behavior. Currently this works via golang equality and compares an entire array against an entire other array as a single value. I expect this to change in the future, but we just don't need it yet.