Unify heartbeat jobs and tasks for simplicity
In heartbeat we have separate notions of jobs and tasks.

This can make following the code and sharing the code among monitors confusing. This patch unifies the two. There is a tiny amount of extra execution overhead.

Tasks only passed a single `common.MapStr`, where we now pass a `*beat.Event`, however, the benefit is that we have a more consistent and simple codebase.